requirements.txt
pywebview
sqlalchemy
appdirs
py7zr
patool